Move testutils/configtestutils to configtest
